Fixed style-home-00 7 pixel jog layout and background color problems.
Reason wrapped most of the content blocks in a div.container which was not in the static HTML templates.  div.container has side margins of 7 pixels and a width of 950 pixels.  I overrode these values for this page and style combination with a selector of higher specificity.

Pretty much the same story for the background color.