CSS Oversized Top Image Fix
Make maximum width of top image (#imagetopframe img) 450 pixels.  This way one large image are shrunk and small images are not enlarged.

Explicitly set top image frame (#imagetopframe) to 450 pixels width and set overflow to hidden.  This way the image is still cropped if the browser does not support max-width.